Refactor JobsViewUpdater implementation into cpp file
As part of doing this, it allows for the anonymization of the
HIDDEN_LOCAL namespace in #26625. However, the code there is totally
duplicated anyway, so there may be scope for refactoring out so that any
changes are used in both classes.

RE #26539
